Splitting Shared framework into SPM
