+++ Sources/HANetworking/Sources/SecurityExceptions.swift
